project/requirements.txt
streamlit
transformers
torch
torchvision
opencv-python
Pillow
numpy
timm
hf_xet